requirements.txt
langchain==1.2.11
langchain-community==0.4.1
langchain-openai==1.1.11
langchain-text-splitters==1.1.1
openai==2.26.0
pinecone==8.1.0
pypdf==6.8.0
python-dotenv==1.2.2
tiktoken==0.12.0
fastapi
uvicorn